Move rsspp::Item into a separate file
